python-backend/requirements.txt
spacy==3.8.11
numpy==1.26.4
scikit-learn
nltk
fastapi
uvicorn
python-multipart
opencv-python-headless
torch
transformers
Pillow
